Padding struct -> MarginPadding (in its own file).
