+++ osu.Framework/osu.Framework.csproj
@@ -117,6 +117,7 @@
     <Compile Include="Graphics\Performance\FpsDisplay.cs" />
     <Compile Include="Graphics\Performance\FrameTimeDisplay.cs" />
     <Compile Include="Graphics\Performance\PerformanceOverlay.cs" />
+    <Compile Include="Graphics\Primitives\MarginPadding.cs" />
     <Compile Include="Graphics\Sprites\SpriteDrawNode.cs" />
     <Compile Include="Graphics\Transformations\ITransform.cs" />
     <Compile Include="Graphics\Transformations\TransformAlpha.cs" />
